log events for spam, virus
git-svn-id: https://ut.svn.beanstalkapp.com/ngfw/alpine/alpine@1316 8ebf8b4e-06f3-0310-961f-89f74618be54

+++ tran/mail/hier/usr/share/metavize/schema/mail-casing/schema-0.sql
@@ -33,7 +33,7 @@ CREATE TABLE tr_mail_message_info_addr (
 CREATE TABLE tr_mail_message_stats (
     id int8 NOT NULL,
     msg_id int8 NOT NULL,
-    msg_bytes long,
+    msg_bytes int8,
     int msg_attachments,
     PRIMARY KEY (id));
